+++ src/features/AreaCoordinates/CoordinateInput/CoordinateInput.tsx
