+++ src/features/AreaCoordinates/CoordinatesDialog/CoordinatesDialog.tsx
