+++ src/features/AreaCoordinates/ErrorMessage/ErrorMessage.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@equinor/eds-core-react';
-import { CoordinateErrorType } from '../AreaCoordinates/hooks/AreaCoordinates.hooks';
+import { CoordinateErrorType } from '../hooks/AreaCoordinates.hooks';
 
 export const ErrorMessage = ({ errors }: { errors: CoordinateErrorType }) => {
   return (
