+++ src/features/AreaCoordinates/tests/mockedData.ts
