+++ src/features/ModelView/ModelView.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
-import { CoordinatesDialog } from '../../components/AreaCoordinates/CoordinatesDialog/CoordinatesDialog';
 import { ModelMetadataView } from '../../features/ModelView/ModelMetadataView/ModelMetadataView';
 import { DeleteModel } from './DeleteButton/DeleteModel';
 import { ModelAreaCoordinates } from './ModelAreaCoordinates/ModelAreaCoordinates';
 import { ModelFilesView } from './ModelFilesView/ModelFilesView';
 import * as Styled from './ModelView.styled';
+import { CoordinatesDialog } from '../AreaCoordinates/CoordinatesDialog/CoordinatesDialog';
 
 export const ModelView = () => {
   const [open, setOpen] = useState<boolean>(false);
